+++ lib/src/widget/impl/facade.dart
@@ -11,13 +11,27 @@ class CloudFlareTurnstile extends StatelessWidget implements i.CloudFlareTurnsti
   @override
   final String siteKey;
 
-  /// A Turnstile widget options
+  /// A customer value that can be used to differentiate widgets under the
+  /// same sitekey in analytics and which is returned upon validation.
+  ///
+  /// This can only contain up to 32 alphanumeric characters including _ and -.
   @override
-  final TurnstileOptions options;
+  final String? action;
+
+  /// A customer payload that can be used to attach customer data to the
+  /// challenge throughout its issuance and which is returned upon validation.
+  ///
+  /// This can only contain up to 255 alphanumeric characters including _ and -.
+  @override
+  final String? cData;
 
   /// A base url of turnstile Site
   @override
   final String baseUrl;
+
+  /// A Turnstile widget options
+  @override
+  final TurnstileOptions options;
 
   /// A controller for an Turnstile widget
   @override
@@ -73,6 +87,8 @@ class CloudFlareTurnstile extends StatelessWidget implements i.CloudFlareTurnsti
   CloudFlareTurnstile({
     super.key,
     required this.siteKey,
+    this.action,
+    this.cData,
     this.baseUrl = 'http://localhost/',
     TurnstileOptions? options,
     this.controller,